examples in rust idiomatic way
