+++ Cargo.toml
@@ -27,3 +27,14 @@ httpmock = "0.7.0"
 mockall = "0.13.1"
 mockall_double = "0.3.1"
 ureq = "2"
+url = "2"
+http = "1"
+thiserror = "2"
+reqwest = { version = "0", default-features = false, features = [
+  "blocking",
+  "rustls-tls",
+  "rustls-tls-native-roots",
+  "socks",
+] }
+dotenvy = "0"
+
