--- examples/http-client/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "http-client"
-description = "HTTP Client that implements libraries HTTP Client Trait to be used in examples"
-version = "0.0.1"
-edition = "2021"
-authors = ["The New Relic Agent Control Team"]
-publish = false
-license-file = "./../LICENSE.md"
-
-[dependencies]
-nr-auth = { path = "../../" }
-url = "2"
-http = "1"
-thiserror = "2"
-reqwest = { version = "0", default-features = false, features = [
-  "blocking",
-  "rustls-tls",
-  "rustls-tls-native-roots",
-  "socks",
-] }
