--- examples/http-client/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod client;
