--- examples/retrieve-token/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "retrieve-token-example"
-description = "Example to retrieve a token from System Identity Service using the library"
-version = "0.0.1"
-edition = "2021"
-authors = ["The New Relic Agent Control Team"]
-publish = false
-license-file = "./../LICENSE.md"
-
-[dependencies]
-nr-auth = { path = "../../" }
-http-client = {path = "../http-client"}
-url = "2"
-dotenvy = "0"
-
-[[bin]]
-name = "retrieve-token-example"
-path = "src/main.rs"
